login/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib==1.7.4
bcrypt==4.0.1
pyjwt